workflow enchancements

+++ boo.py
@@ -1,30 +1,36 @@
 import pandas as pd
 from columns import shorten, COLUMNS_SHORT, DTYPES_SHORT
 from file import curl, yield_rows, save_rows
-from paths import raw, processed
+from path import raw, processed, canonical
 from year import make_url
 
-# FIXME: must fail of downloading an HTML
+def preclean(path, force: bool):
+    """Delete an exisiting file if *force* flag is set to True"""
+    if force is True and path.exists():
+       path.unlink()
 
-def download(year, force=False):
+# MAYBE: show line count  
+
+def download(year: int, force=False):
     """Download file from Rosstat."""
     path, url = raw(year), make_url(year), 
-    if force:
-       path.unlink()
+    preclean(path, force)
     print("Downloading", url)
     curl(path, url)
     print("Saved as", path)
-    return path       
-       
 
-def cut(year, worker=shorten, column_names=COLUMNS_SHORT, force=False):
-    """Create local file with fewer columns. 
-       Columns in new file will be named according to *COLUMNS_SHORT*.
-       Rows will be modified with *worker* function.
+     
+def cut(year: int, force=False):
+    cut_columns(year, force=force)
+
+
+def cut_columns(year, worker=shorten, column_names=COLUMNS_SHORT, force=False):
+    """Create smaller local file with fewer columns. 
+       *worker* function trims columns.
+       *COLUMNS_SHORT* are outcoming column names.
     """    
     src, dst = raw(year), processed(year)
-    if force:
-       dst.unlink()
+    preclean(dst, force)
     print("Reading from", dst)    
     print("Saving to", dst)    
     save_rows(path=dst, 
@@ -36,13 +42,32 @@ def cut(year, worker=shorten, column_names=COLUMNS_SHORT, force=False):
 def dataframe(path, dtypes):
     with open(path, 'r', encoding='utf-8') as f:
         return pd.read_csv(f, dtype=dtypes)  
-
     
-def read_processed_df(year):
+    
+def read_intermediate_df(year: int):
     src = processed(year)    
     return dataframe(src, DTYPES_SHORT)
 
-# FIXME: move parsing logic here
 
-def read_canonical_df(year):
-    pass
+def put(year: int, force=False):
+    make_canonical_df(year, force)
+    
+    
+def make_canonical_df(year: int, worker, column_names, force=False):
+    df = read_intermediate_df(year)    
+    # TODO: add parsing logic here from row.py
+    # MAYBE: save dtypes as json, use them if available to speed up df import  
+    dst = canonical(year)
+    pass   
+
+
+def read_df(year):
+    src = canonical(year)
+    # TODO: use types list
+    return dataframe(src)
+
+
+def acquire(year: int):
+    download(year)
+    cut(year)
+    put(year)
